refactor(layout): colocate span placement lowering
