+++ packages/gofish-graphics/src/ast/constraints/placementSolver.ts
@@ -15,9 +15,9 @@ import type { NestConstraint } from "./nest";
 import { lowerNestPlacement } from "./nest";
 import type { PositionConstraint } from "./position";
 import { lowerPositionPlacement } from "./position";
-import type { SpanConstraint } from "./span";
+import type { SpanConstraint, SpanExtent } from "./span";
+import { collectSpanExtents, lowerSpanPlacements } from "./span";
 import type { Axis, AlignAnchor, ConstraintPosScales } from "./shared";
-import { BBox, type BBoxFacet } from "./bbox";
 import type {
   NodeId,
   PlacementFactEmitter,
@@ -44,27 +44,6 @@ type PlacementConstraint =
   | SpanConstraint
   | NestConstraint
   | GridConstraint;
-
-/** One emitted span extent: the target owns both edges on one axis. */
-interface SpanPlacement {
-  name: string;
-  target: Placeable;
-  axis: Axis;
-  owned: { min: number; max: number };
-  owner: string;
-}
-
-interface SpanGroup {
-  name: string;
-  target: Placeable;
-  axis: Axis;
-  bbox: BBox;
-  owner: string;
-}
-
-interface SpanExtent extends PlacementSpan {
-  size: number;
-}
 
 export interface LoweredPlacement {
   program: PlacementProgram;
@@ -137,95 +116,6 @@ function resolveCoordinate(
   scale: ((value: number) => number) | undefined
 ): number | undefined {
   return compilePlacementCoordinate(coordinate, scale);
-}
-
-function emitSpanPlacements(
-  constraint: SpanConstraint,
-  targets: Map<string, Placeable>,
-  posScales: ConstraintPosScales | undefined,
-  owner: string
-): SpanPlacement[] {
-  const out: SpanPlacement[] = [];
-  const emitAxis = (
-    axis: Axis,
-    span: [MaybeValue<number>, MaybeValue<number>] | undefined
-  ) => {
-    if (span === undefined) return;
-    const min = resolveCoordinate(span[0], posScales?.[axisIndex(axis)]);
-    const max = resolveCoordinate(span[1], posScales?.[axisIndex(axis)]);
-    if (min === undefined || max === undefined) return;
-    for (const child of constraint.children) {
-      const target = targets.get(child.name);
-      if (target)
-        out.push({
-          name: child.name,
-          target,
-          axis,
-          owned: { min, max },
-          owner,
-        });
-    }
-  };
-  emitAxis("x", constraint.x);
-  emitAxis("y", constraint.y);
-  return out;
-}
-
-function collectSpanExtents(placements: SpanPlacement[]): SpanExtent[] {
-  const byTarget = new Map<Placeable, [SpanGroup?, SpanGroup?]>();
-  const groups: SpanGroup[] = [];
-  for (const placement of placements) {
-    const idx = axisIndex(placement.axis);
-    let axes = byTarget.get(placement.target);
-    if (!axes) {
-      axes = [undefined, undefined];
-      byTarget.set(placement.target, axes);
-    }
-    let group = axes[idx];
-    if (!group) {
-      group = {
-        name: placement.name,
-        target: placement.target,
-        axis: placement.axis,
-        bbox: new BBox(),
-        owner: placement.owner,
-      };
-      axes[idx] = group;
-      groups.push(group);
-    }
-
-    for (const [facet, value] of Object.entries(placement.owned) as [
-      BBoxFacet,
-      number,
-    ][]) {
-      const conflict = group.bbox.add(facet, value, placement.owner);
-      if (conflict) {
-        throw new Error(
-          `Constraint span conflict on ${placement.axis}: ${conflict.owner} ` +
-            `asserts ${conflict.facet}=${conflict.asserted}, but ` +
-            `${conflict.priorOwner} implies ${conflict.implied}`
-        );
-      }
-    }
-  }
-
-  return groups.flatMap((group) => {
-    const min = group.bbox.read("min");
-    const max = group.bbox.read("max");
-    if (min === undefined || max === undefined) return [];
-    const size = max - min;
-    return [
-      {
-        type: "span",
-        name: group.name,
-        axis: group.axis,
-        min,
-        max,
-        size,
-        owner: group.owner,
-      },
-    ];
-  });
 }
 
 function compareRank(a: PlacementWeakPin, b: PlacementWeakPin): number {
@@ -493,11 +383,12 @@ export function lowerPlacementConstraints(
 
   const spanPlacements = constraints.flatMap((constraint, constraintIndex) =>
     constraint.type === "span"
-      ? emitSpanPlacements(
+      ? lowerSpanPlacements(
           constraint,
           targets,
-          posScales,
-          `span[${constraintIndex}]`
+          `span[${constraintIndex}]`,
+          (axis, coordinate) =>
+            resolveCoordinate(coordinate, posScales?.[axisIndex(axis)])
         )
       : []
   );
